Added Checks that ensure if data is present in All Services.

+++ src/main/java/com/project/codemic/Codemic/service/ActivityServiceImpl.java
@@ -44,6 +44,8 @@ public class ActivityServiceImpl implements ActivityService {
 
     @Override
     public void deleteActivity(Integer id) {
-        activityRepository.deleteById(id);
+        Optional<Activity> activityOptional = activityRepository.findById(id);
+        if (activityOptional.isPresent())
+            activityRepository.deleteById(id);
     }
 }
